+++ CHANGELOG.md
@@ -19,3 +19,4 @@ All changes for this project will be registered here
 ## [1.0.3] - 2018-03-22
 ### Changed
 - assets/js/simple-dual-listbox.js addRemoveElement() function modified to support sort of available items
+- SimpleDualListbox.php Static variables for sort types